refactor: refactor the entry file name
